Use user timezones, not offset
Right now a seconds of offset is passed in the URL. This is problematic during
daylight savings time, as the offset changes for constant locations.

Rather than trying to calculate an offset client-side and pass it in the URL,
move this server side. This allows URLs to be pinned to a physical location.