Fix fetching deadlock when connection issues occur
When there's network issues (thanks vodafone), chunks can fail to fetch. When this
happens our calculation to reset the timer based on fetched data returns 0, which
deadlocks the ticker (https://github.com/golang/go/issues/49315). As this will panic
in future releases as well, reset the ticker to 5 seconds when this occurs. This should
stop hammering things when network issues occur too. Exponential backoff would be nice,
but y'know.

Also upgrade to go 1.18rc while here, because why not.

Fixes #5